refactor: move api calls School Page to SchoolAPI.ts
